refactor: split file in 2
